requirements.txt
streamlit
pandas
matplotlib
scikit-learn
python-dotenv
openai
